Update patch set 1
Patch Set 1: No score

The reason SoundRecorder isn't exposed in the launcher is because it's lacking some features, such as a basic ability to manage its own recordings. Before making SoundRecorder a first class citizen in the launcher, I would like to see such functionality added, i.e. the user should be able to view, play and delete their existing recordings (it doesn't count that you can currently do this in the music app, it needs to be accessible from SoundRecorder)

Patch-set: 1