Fix a Window Leak issue caused by AlertDialog.
Issue Description:
Record a period sound, then play it with SoundRecorder, press "return"
key twice during playing. Log shows there is a window leak.

Root Cause:
At first time of pressing "return" key, the recorded file would be saved
to media DB by calling saveSample(), but at the second time of pressing
"return" key, SoundRecorder would try to save again but certainly failed.
Then it tries to draw an AlertDialg to warn users, but it would invoke
finish() before AlertDialog showing. In this case, the AlertDialog is
a child of the MainAcitivity, so that WindowManager can not find the
parent, and a Window Leak error jumps out.

Solution:
Move the AlertDialog creating method to callback function onCreateDialog()
so that GC would recycle the AlertDialog automatically when finish().

Change-Id: I83c886a7342a6ac3f9def8b80e525c6bed8fce47
Signed-off-by: gxie <gxie@marvell.com>