Fix npe in soundrecorder
In certain cases when the file system is full, allocation of
cursor might fail. When this happens the call to cursor.close()
will result in a NullPointerException.

Moved close() inside if statement checking that cursor is not
null.

Also added a try/finally pattern to ensure that cursor is always
closed.

Change-Id: I1b351cb18cfbaef2954e20baf3f73abf439a901c