Update patch set 1
Patch Set 1:

(2 comments)

This change does 2 things: move the close() within the if != null. That part looks good.  But it also adds a try. That might be useful [I don't which exceptions cursor operations throw] but it doesn't match the commit message. Please provide more information about which exceptions are likely to occur within the try.

Patch-set: 1
Label: Code-Review=0